+++ projects/ng2-qgrid-theme-material/src/lib/theme.module.ts
@@ -70,8 +70,8 @@ import {
 	VisibilityModule,
 	VscrollModule,
 	MarkupModule,
+	ChangeDetectorModule
 } from 'ng2-qgrid';
-import { ChangeDetectorModule } from '@qgrid/ngx';
 
 @NgModule({
 	declarations: [