+++ projects/ngx-qgrid-plugins/src/lib/pagination/pager.component.ts
@@ -12,10 +12,11 @@ import { PaginationState } from '@qgrid/core/pagination/pagination.state';
 export class PagerComponent implements OnInit, OnChanges {
 	private pagerState = this.stateAccessor.setter(PaginationState);
 
-	context: { $implicit: PagerPlugin };
+	context: { $implicit: PagerPlugin, plugin: PagerComponent };
 
 	@Input('size') set paginationSize(size: number) { this.pagerState({ size }); }
 	@Input('sizeList') set paginationSizeList(sizeList: number[]) { this.pagerState({ sizeList }); }
+	@Input('showCount') showCount = false;
 
 	constructor(
 		private plugin: GridPlugin,
@@ -32,7 +33,7 @@ export class PagerComponent implements OnInit, OnChanges {
 	ngOnInit() {
 		const { model, observe } = this.plugin;
 		const pager = new PagerPlugin(this.plugin);
-		this.context = { $implicit: pager };
+		this.context = { $implicit: pager, plugin: this };
 
 		observe(model.paginationChanged)
 			.subscribe(() => this.cd.detectChanges());