+++ projects/ngx-qgrid/src/public-api.ts
@@ -53,6 +53,7 @@ export { ThemeService } from './lib/theme/theme.service';
 export { ResizeModule } from './lib/resize/resize.module';
 
 export { ChangeDetectorModule } from './lib/change-detector/change-detector.module';
+export { DirtyDirective } from './lib/change-detector/dirty.directive';
 
 export {
 	DomTr,