+++ projects/qgrid-core/grid/grid.host.d.ts
@@ -1,4 +1,4 @@
-import { GridPlugin } from '@qgrid/ngx';
+import { GridPlugin } from '../plugin/grid.plugin';
 
 export declare class GridHost {
 	constructor(