+++ CHANGELOG.md
@@ -50,6 +50,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - *ngIf for column component works.
 - Fix row options icon is jumping on edit.
 - Row resize algorithm.
+- Better paddings for column filter and chooser.
 
 ## [9.0.1] - 2020-03-13
 ### Fixed